jrfjrf

+++ src/app/pages/districts/districts-list/districts-list.component.ts
@@ -3,6 +3,7 @@ import { CommonModule } from '@angular/common';
 import { restApiService } from 'src/app/core/services/rest-api.service copy';
 import { Router } from '@angular/router';
 import Swal from 'sweetalert2';
+import { AuthenticationService } from 'src/app/core/services/auth.service';
 
 @Component({
   selector: 'app-districts-list',
@@ -13,12 +14,17 @@ import Swal from 'sweetalert2';
 })
 export class DistrictsListComponent {
   districts
-  constructor(private rnpService: restApiService,private router: Router) { }
+  constructor(private rnpService: restApiService,private router: Router,private authService: AuthenticationService) { }
 
   ngOnInit(): void {
    this.getReources()
 
 }
+getConnectedUserRole(){
+  if(this.authService.loadToken()){
+    return JSON.parse(atob(this.authService.loadToken().split('.')[1])).roles[0].authority
+  }
+ }
 getReources(){
   this.rnpService.getResourceAll('districts').subscribe(data=>{
     this.districts = data['_embedded'].districts
@@ -30,36 +36,39 @@ addResource(){
 
 }
 onDeleteResource(url:string){
-Swal.fire({
-  title: 'هل أنت متأكد؟',
-  text: 'سوف يتم الحذف بصفة نهائية!',
-  icon: 'warning',
-  showCancelButton: true,
-  confirmButtonColor: '#364574',
-  cancelButtonColor: 'rgb(243, 78, 78)',
-  cancelButtonText: 'إلغاء',
-  customClass:{
-    title: 'kuffi',
-    confirmButton: 'kuffi',
-    container: 'kuffi'
-  },
-  confirmButtonText: 'حذف'
-}).then(result => {
- 
-  if (result.value) {
-    this.rnpService.deleteResource('districts',url).subscribe(data=>{
-      this.getReources()
-      Swal.fire({text:'لقد تم حذف الدائرة', confirmButtonColor: '#364574',   customClass:{
+  if(this.getConnectedUserRole()!='USER-AAL'){
+    Swal.fire({
+      title: 'هل أنت متأكد؟',
+      text: 'سوف يتم الحذف بصفة نهائية!',
+      icon: 'warning',
+      showCancelButton: true,
+      confirmButtonColor: '#364574',
+      cancelButtonColor: 'rgb(243, 78, 78)',
+      cancelButtonText: 'إلغاء',
+      customClass:{
         title: 'kuffi',
         confirmButton: 'kuffi',
         container: 'kuffi'
-      }, icon: 'success',});
-       },err=>{
-        this.modelError('لا يمكن حدف الدائرة')
-       })
-
+      },
+      confirmButtonText: 'حذف'
+    }).then(result => {
+     
+      if (result.value) {
+        this.rnpService.deleteResource('districts',url).subscribe(data=>{
+          this.getReources()
+          Swal.fire({text:'لقد تم حذف الدائرة', confirmButtonColor: '#364574',   customClass:{
+            title: 'kuffi',
+            confirmButton: 'kuffi',
+            container: 'kuffi'
+          }, icon: 'success',});
+           },err=>{
+            this.modelError('لا يمكن حدف الدائرة')
+           })
+    
+      }
+    });
   }
-});
+
  
 }
 onEditResource(p:any){
